+++ openbb_platform/extensions/famafrench/integration/__init__.py
@@ -1,0 +1,1 @@
+"""OpenBB FamaFrench Integration Tests module."""
